requirements.txt
# Discord bot
discord.py>=2.3.0
discord-py-slash-command>=4.2.1

# Web framework
flask>=3.0.0
flask-login>=0.6.3
flask-limiter>=3.5.0
flask-socketio>=5.3.6
flask-wtf>=1.2.1

# WSGI server
gunicorn>=21.2.0
gevent>=23.9.1

# Database
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis
redis>=5.0.0

# Auth / HTTP
authlib>=1.3.0
requests>=2.31.0
httpx>=0.27.0
PyJWT>=2.8.0

# Background worker
schedule>=1.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
structlog>=24.1.0

# Security
werkzeug>=3.0.0
itsdangerous>=2.1.0
cryptography>=42.0.0
